added class Document to antiweb_lib package
