+++ documentation/source/directives.rst
@@ -1,4 +1,3 @@
-""""
 .. _Directives:
 
 **********
